+++ tests/pretty_urls.test.ts
@@ -9,13 +9,3 @@ Deno.test("Disabled pretty URLs", async (t) => {
   await build(site);
   await assertSiteSnapshot(t, site);
 });
-
-Deno.test("Pretty URLs with no-html-extension", async (t) => {
-  const site = getSite({
-    src: "simple",
-    prettyUrls: "no-html-extension",
-  });
-
-  await build(site);
-  await assertSiteSnapshot(t, site);
-});